chore: move provider to new file
